# Reduce memory usage by only caching the metadata of Secret resources