@maelvls 2024-07-11
Hey, well done with the benchmarks on this PR. Super happy to see more data-driven approaches like this one. And thank you for providing the detailed instructions for reproducing your experiments. 👏

> To reduce the load on the K8S API server when cainjector starts up, caused by the initial listing of Secret resources in the cluster.

Are there any downsides to disabling caching? When using Upbound's Crossplane, at least [900 CRDs](https://blog.crossplane.io/crd-scaling-provider-families/) are installed and are expected to be CA-injected by cert-manager's cainjector. Would disabling be less favorable in that case? I imagine that on startup all the secrets will need to be listed to check that they match the contents of the CA injected in the CRDs.

> 1000 RSA 2048 Secrets amounts to ~6MB of data so the difference in memory usage is not dramatic.

I think the improvement would be much more dramatic in a cluster with lots of large Helm release secrets hanging around. This is something you will often see in prod clusters. 

> kubectl get secret -l controller.cert-manager.io/fao -A -o json | dd bs=1M of=/dev/null

Are you sure the `controller.cert-manager.io/fao` annotation trick was implemented in the cert-manager cainjector? I think it was only implemented in the cert-manager controller.

I haven't reviewed the code yet, I'll come back to it later.