@wallrj 2024-07-12
@maelvls I added some example curl commands to the description showing the difference in response size when you send the PartialObjectMetadataList  content negotiation header.