@maelvls 2024-07-12
Richard presented this PR to [yesterday's dev biweekly](https://docs.google.com/document/d/1Tc5t6ylY9dhXAan1OjOoldeaoys1Yh4Ir710ATfBa5U/edit#bookmark=kix.o7513fuynfby), some notes:

- Richard made me aware of the `accept` header described in [2334-graduate-server-side-get-and-partial-objects-to-GA.md](https://github.com/kubernetes/enhancements/tree/master/keps/sig-api-machinery/2334-graduate-server-side-get-and-partial-objects-to-GA):

  ```
  accept: application/json;as=PartialObjectMetadata;g=meta.k8s.io;v=v1
  ```

  which means cainjector does list all the secrets during the initial "list" HTTP call, but it doesn't have to process the contents of these secrets; only their metadata.
- On the topic of the downsides of disabling caching, I agree that even with over 1000 CRDs, the decrease in memory usage is much preferable over having a cache. Caching would only be useful in case we were accessing or listing the secrets often, which we realistically don't do.