@wallrj 2024-07-12
Thanks @maelvls 

I'll fix that nit in another PR to save another round of e2e tests and review.

/unhold